Check if model allows videos before deleting in AdminModelObserver

+++ src/Observers/AdminModelObserver.php
@@ -79,7 +79,9 @@ class AdminModelObserver
      */
     public function deleted(AdminModel $model)
     {
-        $model->videos()->delete();
+        if ($model->allowsVideo()) {
+            $model->videos()->delete();
+        }
     }
 
     /**
